==> 01_json_scraping.ipynb <==
import necessary module

import json

Give the file path and open the file 

file_path = "../../data/json_handling/spotify_playlist.json"

with open(file_path, "r") as json_file:
    data = json.load(json_file)

data

for track in data["contents"]["items"]:
    print(track["playCount"])
    print(track["name"])
    print()
